requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
antlr4-python2-runtime=4.9.3=pypi_0
appnope=0.1.2=py39hecd8cb5_1001
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py39hca72f7f_0
asgiref=3.4.1=pyhd3eb1b0_0
attrs=21.4.0=pyhd3eb1b0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.10.0=pyh06a4308_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blosc=1.21.0=h2842e9f_0
bottleneck=1.3.2=py39he3068b8_1
brotli=1.0.9=hb1e8313_2
brotlipy=0.7.0=py39h9ed2024_1003
brunsli=0.1=h23ab428_0
bs4=4.10.0=hd3eb1b0_0
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=hca72f7f_0
ca-certificates=2022.3.29=hecd8cb5_1
certifi=2021.10.8=py39hecd8cb5_2
cffi=1.15.0=py39hc55c11b_1
cfitsio=3.470=hee0f690_6
charls=2.2.0=h23ab428_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39hecd8cb5_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.4=pyh9f0ad1d_0
conda=4.12.0=py39hecd8cb5_0
conda-package-handling=1.7.3=py39h9ed2024_1
configparser=5.2.0=pyhd8ed1ab_0
crayons=0.4.0=pyh9f0ad1d_0
cryptography=36.0.0=py39hf6deb26_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.11.0=py39h9ed2024_0
dask-core=2021.10.0=pyhd3eb1b0_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py39he9d5cce_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
django=3.2.5=pyhd3eb1b0_0
django-pandas=0.6.6=pyhd8ed1ab_0
entrypoints=0.3=py39hecd8cb5_0
ffmpeg=4.2.2=h97e5cf8_0
filelock=3.4.2=pyhd3eb1b0_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.11.0=hd8bbffd_0
fsspec=2022.1.0=pyhd3eb1b0_0
future=0.18.2=py39hecd8cb5_1
gettext=0.21.0=h7535e17_0
giflib=5.2.1=haf1e3a3_0
gmp=6.2.1=h23ab428_2
gnutls=3.6.15=hed9c0bf_0
huggingface_hub=0.2.1=pyhd3eb1b0_0
icu=58.2=h0a44026_3
idna=3.3=pyhd3eb1b0_0
imagecodecs=2021.8.26=py39ha952a84_0
imageio=2.9.0=pyhd3eb1b0_0
importlib-metadata=4.8.2=py39hecd8cb5_0
importlib_metadata=4.8.2=hd3eb1b0_0
intel-openmp=2021.4.0=hecd8cb5_3538
ipykernel=6.9.0=py39h71a6800_0
ipython=7.31.1=py39hecd8cb5_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.5=pyhd8ed1ab_0
jedi=0.18.1=py39hecd8cb5_1
jinja2=3.0.2=pyhd3eb1b0_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9d=h9ed2024_0
jsonschema=3.2.0=pyhd3eb1b0_2
jupyter_client=7.1.2=pyhd3eb1b0_0
jupyter_core=4.9.1=py39hecd8cb5_0
jupyterlab_pygments=0.1.2=py_0
jupyterlab_widgets=1.0.0=pyhd3eb1b0_1
jxrlib=1.1=haf1e3a3_2
kiwisolver=1.3.2=py39he9d5cce_0
krb5=1.19.2=hcd88c3b_0
lame=3.100=h1de35cc_0
lcms2=2.12=hf1fd2bf_0
lerc=3.0=he9d5cce_0
libaec=1.0.4=hb1e8313_1
libcurl=7.80.0=h6dfd666_0
libcxx=12.0.0=h2f01273_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20210910=hca72f7f_0
libev=4.33=h9ed2024_1
libffi=3.3=hb1e8313_2
libgfortran=3.0.1=h93005f0_2
libiconv=1.16=h1de35cc_0
libidn2=2.3.2=h9ed2024_0
libnghttp2=1.46.0=ha29bfda_0
libopus=1.3.1=h1de35cc_0
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_1
libprotobuf=3.19.1=h8346a28_0
libsodium=1.0.18=h1de35cc_0
libssh2=1.9.0=ha12b0ac_1
libtasn1=4.16.0=h9ed2024_0
libtiff=4.2.0=h87d7836_0
libunistring=0.9.10=h9ed2024_0
libuv=1.40.0=haf1e3a3_0
libvpx=1.7.0=h378b8a2_0
libwebp=1.2.2=h56c3ce4_0
libwebp-base=1.2.2=hca72f7f_0
libxml2=2.9.12=hcdb78fc_0
libxslt=1.1.34=h83b36ba_0
libzopfli=1.0.3=hb1e8313_0
llvm-openmp=12.0.0=h0dcd299_1
locket=0.2.1=py39hecd8cb5_2
lxml=4.8.0=py39h65b224f_0
lz4-c=1.9.3=h23ab428_1
markupsafe=2.0.1=py39h9ed2024_0
matplotlib=3.5.0=py39hecd8cb5_0
matplotlib-base=3.5.0=py39h4f681db_0
matplotlib-inline=0.1.2=pyhd3eb1b0_2
mistune=0.8.4=py39h9ed2024_1000
mkl=2021.4.0=hecd8cb5_637
mkl-service=2.4.0=py39h9ed2024_0
mkl_fft=1.3.1=py39h4ab4a9b_0
mkl_random=1.2.2=py39hb2f4e1b_0
munkres=1.1.4=py_0
nbclient=0.5.11=pyhd3eb1b0_0
nbconvert=6.3.0=py39hecd8cb5_0
nbformat=5.1.3=pyhd3eb1b0_0
ncurses=6.3=hca72f7f_2
nest-asyncio=1.5.1=pyhd3eb1b0_0
nettle=3.7.3=h230ac6f_1
networkx=2.6.3=pyhd3eb1b0_0
ninja=1.10.2=py39haf03e11_3
nltk=3.7=pyhd3eb1b0_0
notebook=6.4.8=py39hecd8cb5_0
numexpr=2.8.1=py39h2e5f0a9_0
numpy=1.21.2=py39h4b4dc7a_0
numpy-base=1.21.2=py39he0bd621_0
openh264=2.1.1=h8346a28_0
openjpeg=2.4.0=h66ea3da_0
openssl=1.1.1n=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.2=py39he9d5cce_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
partd=1.2.0=pyhd3eb1b0_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.0.1=py39hde71d04_0
prometheus_client=0.13.1=pyhd3eb1b0_0
prompt-toolkit=3.0.20=pyhd3eb1b0_0
protobuf=3.19.1=py39he9d5cce_0
psycopg2=2.8.6=py39hbcfaee0_1
ptyprocess=0.7.0=pyhd3eb1b0_2
pycosat=0.6.3=py39h9ed2024_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyrsistent=0.18.0=py39hca72f7f_0
pysocks=1.7.1=py39hecd8cb5_0
python=3.9.5=h88f2d9e_3
python-dateutil=2.8.2=pyhd3eb1b0_0
python.app=3=py39h9ed2024_0
python_abi=3.9=2_cp39
pytorch=1.11.0=py3.9_0
pytz=2021.3=pyhd3eb1b0_0
pywavelets=1.1.1=py39he3068b8_4
pyyaml=6.0=py39hca72f7f_1
pyzmq=22.3.0=py39he9d5cce_2
readline=8.1.2=hca72f7f_1
regex=2021.11.2=py39hca72f7f_0
requests=2.27.1=pyhd3eb1b0_0
ruamel_yaml=0.15.100=py39h9ed2024_0
sacremoses=0.0.43=pyhd3eb1b0_0
scikit-image=0.18.3=py39hae1ba45_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.7.3=py39h8c7af03_0
seaborn=0.11.2=pyhd3eb1b0_0
selenium=3.141.0=py39h9ed2024_1000
send2trash=1.8.0=pyhd3eb1b0_1
sentence-transformers=2.2.0=pyhd8ed1ab_0
sentencepiece=0.1.95=py39haf03e11_0
setuptools=58.0.4=py39hecd8cb5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.8=hb1e8313_0
soupsieve=2.3.1=pyhd3eb1b0_0
sqlite=3.37.2=h707629a_0
sqlparse=0.4.1=py_0
terminado=0.13.1=py39hecd8cb5_0
testpath=0.5.0=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.11=h7bc2e8c_0
tokenizers=0.10.3=py39h7bafbf5_1
toolz=0.11.2=pyhd3eb1b0_0
torchvision=0.12.0=py39_cpu
tornado=6.1=py39h9ed2024_0
tqdm=4.62.3=pyhd3eb1b0_1
traitlets=5.1.1=pyhd3eb1b0_0
transformers=4.11.3=py_0
typing-extensions=3.10.0.2=hd3eb1b0_0
typing_extensions=3.10.0.2=pyh06a4308_0
tzdata=2021e=hda174b7_0
urllib3=1.26.8=pyhd3eb1b0_0
vue=2.6.11=0
wcwidth=0.2.5=pyhd3eb1b0_0
webdriver-manager=3.5.3=pyhd8ed1ab_0
webencodings=0.5.1=py39hecd8cb5_1
widgetsnbextension=3.5.2=py39hecd8cb5_0
x264=1!157.20191217=h1de35cc_0
xz=5.2.5=h1de35cc_0
yaml=0.2.5=haf1e3a3_0
zeromq=4.3.4=h23ab428_0
zfp=0.5.5=he9d5cce_6
zipp=3.7.0=pyhd3eb1b0_0
zlib=1.2.11=h4dc903c_4
zstd=1.4.9=h322a384_0
